Unite stable based on ionic RC3 with updated folder structure
