--- app/app.html
@@ -1,17 +1,0 @@
-<ion-menu [content]="content">
-
-  <ion-toolbar>
-    <ion-title>Pages</ion-title>
-  </ion-toolbar>
-
-  <ion-content>
-    <ion-list>
-      <button menuClose ion-item *ngFor="#p of pages" (click)="openPage(p)">
-        {{p.title}}
-      </button>
-    </ion-list>
-  </ion-content>
-
-</ion-menu>
-
-<ion-nav id="nav" [root]="rootPage" #content swipeBackEnabled="false"></ion-nav>
