--- app/pages/addmenu/addmenu.html
@@ -1,42 +1,0 @@
-<ion-navbar *navbar>
-    <button menuToggle>
-        <ion-icon name="menu"></ion-icon>
-    </button>
-    <ion-title>Add Menu</ion-title>
-</ion-navbar>
-
-<ion-content padding class="login">
-    <ion-list>
-        <form #addmenuForm="ngForm" novalidate>
-            <ion-item>
-                <ion-label floating primary>Menu Name</ion-label>
-                <ion-input [(ngModel)]="addmenu.menuname" ngControl="menuname" type="text" #menuname="ngForm" required>
-                </ion-input>
-            </ion-item>
-            <p [hidden]="menuname.valid || submitted == false" danger padding-left>
-                Menu Name is required
-            </p>
-            <ion-item>
-                <ion-label floating primary>Component Name</ion-label>
-                <ion-input [(ngModel)]="addmenu.component" ngControl="component" type="text" #component="ngForm" required>
-                </ion-input>
-            </ion-item>
-            <p [hidden]="component.valid || submitted == false" danger padding-left>
-                Component is required
-            </p>
-            <ion-row responsive-sm>
-                <ion-col>
-                    <button *ngIf="!updateButton" (click)="addMenu(addmenuForm)" type="submit" primary block>Submit</button>
-                    <button *ngIf="updateButton" (click)="updateMenu(addmenuForm)" type="submit" primary block>Update</button>
-                </ion-col>
-            </ion-row>
-        </form>
-    </ion-list>
-    <ion-list>
-        <ion-item *ngFor="#item of pages; #i=index">
-            {{item.title}}
-            <ion-icon class="icon-padding-right" name="remove-circle" item-right (click)="removeMenu(i)"></ion-icon>
-            <ion-icon name="create" item-right (click)="editMenu(i,item)"></ion-icon>
-        </ion-item>
-    </ion-list>
-</ion-content>
