--- app/pages/details/details.html
@@ -1,10 +1,0 @@
-<ion-navbar *navbar>
-    <button menuToggle *ngIf="!selectedItem">
-        <ion-icon name="menu"></ion-icon>
-    </button>
-    <ion-title>Title</ion-title>
-</ion-navbar>
-
-<ion-content>
-    Ashwin
-</ion-content>
