--- app/pages/getting-started/getting-started.scss
@@ -1,11 +1,0 @@
-.home {
-    .image {
-        width: 75px;
-        height: 75px;
-        background-size: cover;
-        background-position: center;
-        padding: 5px 5px 5px 5px; 
-        background-color: white;
-        box-shadow: 0 1px 1px rgba(0,0,0,.04);
-    }
-}
