--- app/pages/list/list.html
@@ -1,30 +1,0 @@
-<ion-navbar *navbar>
-	<button menuToggle>
-    <ion-icon name="menu"></ion-icon>
-  </button>
-	<ion-title>Lists</ion-title>
-</ion-navbar>
-<ion-toolbar>
-  <ion-searchbar 
-  				       *ngIf="enableSearchbar"
-                 [(ngModel)]="searchQuery"
-                 (search)="getItems($event)"
-				 				(clear)="clearSearch()"
-                 placeholder="Search">
-  </ion-searchbar>
-</ion-toolbar>
-<!--<ion-searchbar *ngIf="enableSearchbar" [(ngModel)]="searchQuery" (input)="getItems($event)"></ion-searchbar>-->
-<ion-content>
-	<ion-list>
-		<ion-refresher *ngIf="enablePullToRefresh" (refresh)="doRefresh($event)">
-			<ion-refresher-content></ion-refresher-content>
-		</ion-refresher>
-		<ion-item *ngFor="#item of items" (click)="itemTapped($event, item)">
-			<h2>{{item.title}}</h2>
-			<p [innerHTML]="item.introtext"></p>
-		</ion-item>
-	</ion-list>
-	<ion-infinite-scroll  *ngIf="enableifinitescroll" (infinite)="loadData($event)" threshold="10%">
-		<ion-infinite-scroll-content></ion-infinite-scroll-content>
-	</ion-infinite-scroll>
-</ion-content>
