--- app/pages/list/list.scss
@@ -1,3 +1,0 @@
-.toast-message { 
-    // background: darkgrey;
-}
