--- app/theme/app.md.scss
@@ -1,31 +1,0 @@
-// http://ionicframework.com/docs/v2/theming/
-
-
-// App Shared Variables
-// --------------------------------------------------
-// Shared Sass variables go in the app.variables.scss file
-@import 'app.variables';
-
-
-// App Material Design Variables
-// --------------------------------------------------
-// Material Design only Sass variables can go here
-
-
-// Ionic Material Design Sass
-// --------------------------------------------------
-// Custom App variables must be declared before importing Ionic.
-// Ionic will use its default values when a custom variable isn't provided.
-@import "ionic.md";
-
-
-// App Shared Sass
-// --------------------------------------------------
-// All Sass files that make up this app goes into the app.core.scss file.
-// For simpler CSS overrides, custom app CSS must come after Ionic's CSS.
-@import 'app.core';
-
-
-// App Material Design Only Sass
-// --------------------------------------------------
-// CSS that should only apply to the Material Design app
