+++ src/pages/details/details.html
@@ -1,0 +1,21 @@
+
+<ion-header>
+  <ion-navbar color="primary" >
+	  <button ion-button menuToggle>
+      <ion-icon name="menu"></ion-icon>
+    </button>
+    <ion-title>Details Page</ion-title>
+  </ion-navbar>
+</ion-header>
+
+<ion-content>
+	<ion-refresher (ionRefresh)="doRefresh($event)">
+			<ion-refresher-content refreshingSpinner="circles"></ion-refresher-content>
+	</ion-refresher>
+	<ion-card-content>
+    	<p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
+  	</ion-card-content>												   
+</ion-content>
+
+	
+
