+++ src/pages/details/details.scss
@@ -1,0 +1,2 @@
+page-details {
+}
