+++ src/pages/list/list.html
@@ -1,0 +1,37 @@
+
+<ion-header >
+	<ion-navbar color="primary" >
+	<button ion-button menuToggle>
+		<ion-icon name="menu"></ion-icon>
+	</button>
+	<ion-title>List</ion-title>
+	</ion-navbar>
+</ion-header>
+
+<ion-content>
+	<ion-searchbar
+		*ngIf="enableSearchbar"
+		[(ngModel)]="searchQuery"
+		(search)="getItems($event)"
+		(ionClear)="clearSearch()"
+		placeholder="Search">
+	</ion-searchbar>	
+
+	<ion-refresher *ngIf="enablePullToRefresh" (ionRefresh)="doRefresh($event)">
+		<ion-refresher-content refreshingSpinner="circles"></ion-refresher-content>
+	</ion-refresher>
+
+	<ion-card *ngFor="let item of items" (click)="itemTapped($event, item)">	
+		<img src="{{item.avatar}}" fallbackSrc/>
+		<ion-card-content>
+			<ion-card-title>
+				{{item.title}}
+			</ion-card-title>
+		</ion-card-content>
+	</ion-card> 
+
+	<ion-infinite-scroll  *ngIf="enableifinitescroll" (ionInfinite)="loadData($event)" threshold="10%">
+		<ion-infinite-scroll-content></ion-infinite-scroll-content>
+	</ion-infinite-scroll>	
+</ion-content>
+
