+++ src/pages/list/list.scss
@@ -1,0 +1,2 @@
+page-list {
+}
