+++ src/pages/login/login.html
@@ -1,0 +1,34 @@
+
+<ion-header >
+	<ion-navbar color="primary" >
+		<ion-title>Login</ion-title>
+	</ion-navbar>
+</ion-header>
+
+<ion-content padding class="login">
+	<ion-list>
+		<form #loginForm="ngForm">
+			<ion-item>
+				<ion-label floating primary> SITE URL</ion-label>
+				<ion-input [(ngModel)]="login.siteurl" name="siteurl" type="text" required></ion-input>
+			</ion-item>
+
+			<ion-item>
+				<ion-label floating primary> EMAIL</ion-label>
+				<ion-input [(ngModel)]="login.username" name="username" type="text" required></ion-input>
+			</ion-item>
+
+			<ion-item>
+				<ion-label floating primary> PASSWORD </ion-label>
+				<ion-input [(ngModel)]="login.password" name="password" type="password" required></ion-input>
+			</ion-item>
+
+			<ion-row>
+				<ion-col>
+					<button  (click)="onLogin(loginForm)" type="submit" ion-button secondary block>LOGIN </button>
+				</ion-col>                
+			</ion-row>
+		</form>
+	</ion-list>
+</ion-content>
+
