+++ src/pages/login/login.scss
@@ -1,0 +1,2 @@
+page-login {
+}
